FULL PIPELINE START TO END_PO - 3B/training_models_risks/requirements.txt
# Core
pip==25.1
setuptools==72.1.0
wheel==0.45.1

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.39

# PyTorch (CPU by default; replace with cu121/cu124 if GPU)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# PyTorch Geometric (install separately if GPU — see notes)
# torch-scatter
# torch-sparse
# torch-cluster
# torch-spline-conv
torch-geometric==2.6.1

# Data science / ML
numpy==2.1.3
pandas==2.2.3
scipy==1.15.3
scikit-learn==1.6.1
scikit-image==0.25.0
statsmodels==0.14.4
xgboost==3.0.1
numba==0.61.0
numexpr==2.10.1
sympy==1.13.1
imbalanced-learn==0.13.0

# Visualization
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1
hvplot==0.11.3
holoviews==1.20.2
datashader==0.18.0
bokeh==3.6.2
tqdm==4.67.1

# Hugging Face / NLP
transformers==4.45.2
#datasets==3.1.0
tokenizers==0.20.1
accelerate==1.1.1
peft==0.13.0
#trl==0.12.0
huggingface-hub==0.26.3
safetensors==0.5.3

# Graph analytics
networkx==3.4.2
#node2vec==0.4.6

# IO / file handling
openpyxl==3.1.5
pyarrow==19.0.0
tables==3.10.2
pyyaml==6.0.2

# Web / APIs
requests==2.32.3
httpx==0.28.1
google-api-python-client==2.176.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Storage
fsspec==2025.3.2
s3fs==2025.3.2
aiobotocore==2.19.0
aiohttp==3.11.10
aioitertools==0.7.1
aiosignal==1.2.0

# Dashboard / UI
streamlit==1.45.1

# Notebook / Dev
jupyter==1.1.1
ipykernel==6.29.5
notebook==7.3.2
ipywidgets==8.1.5
black==24.10.0
flake8==7.1.1
pylint==3.3.5
mypy==1.14.1
